requirements.txt
numpy
pandas
streamlit
google-cloud-vision
protobuf==3.20.1
openai